get openxr client; more info ->
this fails due to the pinput-based portion requiring CAP_SYS_PTRACE, which causes the application to run in Secure Execution mode. as a result, the enviroment variables are stripped and OpenXR thus can not find the runtime. so, if we run with CAP_SYS_PTRACE, OpenXR fails, but if we run without it, the PICO-8 interface fails. what can we do here? two applications?